Move tests to shapely proper
