+++ shapely-core/src/tests.rs
@@ -1,67 +1,3 @@
-use crate::{Def, OpaqueConst, Poke, Shape, Shapely, StructDef, mini_typeid};
-
-// Allow dead code in test modules since we're not constructing all enum variants
-#[allow(dead_code)]
-#[derive(Debug, PartialEq, Eq)]
-struct FooBar {
-    foo: u64,
-    bar: String,
-}
-
-impl Shapely for FooBar {
-    fn shape() -> crate::Shape {
-        Shape {
-            typeid: mini_typeid::of::<Self>(),
-            layout: std::alloc::Layout::new::<Self>(),
-            vtable: || crate::ValueVTable {
-                type_name: |f, _opts| write!(f, "FooBar"),
-                display: None,
-                debug: None,
-                default_in_place: None,
-                eq: None,
-                cmp: None,
-                hash: None,
-                drop_in_place: Some(|ptr| unsafe {
-                    std::ptr::drop_in_place(ptr.as_mut_ptr::<Self>())
-                }),
-                parse: None,
-                try_from: None,
-            },
-            def: Def::Struct(StructDef {
-                fields: crate::struct_fields!(FooBar, (foo, bar)),
-            }),
-        }
-    }
-}
-
-#[test]
-fn build_foobar_through_reflection() {
-    let mut poke = Poke::alloc::<FooBar>().into_struct();
-    poke.set_by_name("foo", OpaqueConst::from_ref(&42u64))
-        .unwrap();
-
-    {
-        let bar = String::from("Hello, World!");
-        poke.set_by_name("bar", OpaqueConst::from_ref(&bar))
-            .unwrap();
-        // bar has been moved out of
-        std::mem::forget(bar);
-    }
-
-    let foo_bar = poke.build::<FooBar>();
-
-    // Verify the fields were set correctly
-    assert_eq!(foo_bar.foo, 42);
-    assert_eq!(foo_bar.bar, "Hello, World!");
-
-    assert_eq!(
-        FooBar {
-            foo: 42,
-            bar: "Hello, World!".to_string()
-        },
-        foo_bar
-    )
-}
 
 // #[test]
 // fn build_u64_through_reflection() {
